fix bug: parity attack for any block.
